error handling

+++ db/db_user.py
@@ -2,6 +2,8 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from db.models import DbUser
 from db.hash import Hash
+from fastapi import HTTPException, status
+
 
 def create_user(db: Session, request: UserBase):
     new_user = DbUser(
@@ -20,11 +22,16 @@ def get_all_users(db: Session):
 
 
 def get_user(id:int, db:Session):
-    return db.query(DbUser).filter(DbUser.id==id).first()
+    user = db.query(DbUser).filter(DbUser.id==id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
+    return user
 
 
 def update_user(db:Session, id:int, request:UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
 
     user.update({
         DbUser.username: request.username,
@@ -37,6 +44,9 @@ def update_user(db:Session, id:int, request:UserBase):
 
 def delete_user(db:Session, id:int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
+
     db.delete(user)
     db.commit()
     return "ok"
